[frontend] FilIndexingConfiguration splitting
