+++ opencti-platform/opencti-front/src/private/components/settings/file_indexing/FileIndexingConfiguration.tsx
@@ -15,54 +15,20 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 
 import React, { FunctionComponent, useEffect } from 'react';
 import EnterpriseEdition from '@components/common/EnterpriseEdition';
-import Typography from '@mui/material/Typography';
 import Grid from '@mui/material/Grid';
-import Paper from '@mui/material/Paper';
-import makeStyles from '@mui/styles/makeStyles';
-import Alert from '@mui/material/Alert';
-import Button from '@mui/material/Button';
-import { PauseOutlined, PlayArrowOutlined } from '@mui/icons-material';
 import { graphql, PreloadedQuery, usePreloadedQuery, useQueryLoader } from 'react-relay';
 import {
   FileIndexingConfigurationFilesMetricsQuery,
 } from '@components/settings/file_indexing/__generated__/FileIndexingConfigurationFilesMetricsQuery.graphql';
+import FileIndexingConfigurationRequirements
+  from '@components/settings/file_indexing/FileIndexingConfigurationRequirements';
+import FileIndexingConfigurationAndImpact from '@components/settings/file_indexing/FileIndexingConfigurationAndImpact';
+import FileIndexingConfigurationInformations
+  from '@components/settings/file_indexing/FileIndexingConfigurationInformations';
 import useEnterpriseEdition from '../../../../utils/hooks/useEnterpriseEdition';
-import { useFormatter } from '../../../../components/i18n';
-import { Theme } from '../../../../components/Theme';
 import useAuth from '../../../../utils/hooks/useAuth';
 import { FILE_INDEX_MANAGER } from '../../../../utils/platformModulesHelper';
 import Loader, { LoaderVariant } from '../../../../components/Loader';
-
-const useStyles = makeStyles<Theme>((theme) => ({
-  paper: {
-    height: '100%',
-    minHeight: '100%',
-    margin: '10px 0 0 0',
-    padding: 20,
-    borderRadius: 6,
-  },
-  gridContainer: {
-    marginBottom: 20,
-  },
-  count: {
-    marginTop: 10,
-    fontSize: 30,
-    color: theme.palette.primary.main,
-    textAlign: 'center',
-  },
-  countText: {
-    textAlign: 'center',
-    marginTop: 5,
-    textTransform: 'uppercase',
-    fontSize: 12,
-    fontWeight: 500,
-    color: theme.palette.mode === 'dark' ? '#ffffff' : '#000000',
-  },
-  button: {
-    marginLeft: theme.spacing(2),
-    marginTop: theme.spacing(2),
-  },
-}));
 
 const fileIndexingConfigurationFilesMetricsQuery = graphql`
   query FileIndexingConfigurationFilesMetricsQuery($mimeTypes: [String]) {
@@ -77,9 +43,9 @@ interface FileIndexingConfigurationComponentProps {
   queryRef: PreloadedQuery<FileIndexingConfigurationFilesMetricsQuery>
 }
 
-const FileIndexingConfigurationComponent: FunctionComponent<FileIndexingConfigurationComponentProps> = ({ queryRef }) => {
-  const { n, t, b } = useFormatter();
-  const classes = useStyles();
+const FileIndexingConfigurationComponent: FunctionComponent<FileIndexingConfigurationComponentProps> = ({
+  queryRef,
+}) => {
   const isEnterpriseEdition = useEnterpriseEdition();
   const { platformModuleHelpers } = useAuth();
   const isModuleWarning = platformModuleHelpers.isModuleWarning(FILE_INDEX_MANAGER);
@@ -88,7 +54,7 @@ const FileIndexingConfigurationComponent: FunctionComponent<FileIndexingConfigur
   const { filesMetrics } = usePreloadedQuery<FileIndexingConfigurationFilesMetricsQuery>(fileIndexingConfigurationFilesMetricsQuery, queryRef);
 
   const totalFiles = filesMetrics?.globalCount;
-  const dataToIndex = filesMetrics?.globalSize; // TODO managing different units of size  14000 => 14 GO  14 => 14 MO
+  const dataToIndex = filesMetrics?.globalSize;
   const indexedFiles = 2;
   const volumeIndexed = 1;
 
@@ -100,106 +66,26 @@ const FileIndexingConfigurationComponent: FunctionComponent<FileIndexingConfigur
       {!isEnterpriseEdition && (
         <EnterpriseEdition />
       )}
-      <Grid container={true} spacing={3} classes={{ container: classes.gridContainer }}>
-        <Grid item={true} xs={12}>
-          <Typography variant="h4" gutterBottom={true}>
-            {t('Requirements')}
-          </Typography>
-          <Paper classes={{ root: classes.paper }} variant="outlined">
-            <Alert
-                classes={{ root: classes.alert, message: classes.message }}
-                severity={isModuleWarning ? 'warning' : 'info'}
-                variant="outlined"
-                style={{ position: 'relative' }}
-            >
-              {t('File indexing needs one of these requirements: ')}
-              <ul>
-                <li>Elasticsearch &gt;= 8.4</li>
-                <li>Elasticsearch &lt; 8.4 with ingest-attachment plugin</li>
-                <li>OpenSearch with ingest-attachment plugin</li>
-              </ul>
-            </Alert>
-          </Paper>
-        </Grid>
-      </Grid>
+      <FileIndexingConfigurationRequirements
+        isModuleWarning={isModuleWarning}
+      />
         {isEnterpriseEdition && !isModuleWarning && (
           <Grid container={true} spacing={3}>
             <Grid item={true} xs={6} style={{ marginTop: 30 }}>
-              <Typography variant="h4" gutterBottom={true}>
-                {t('Configuration and impact')}
-              </Typography>
-              <Paper classes={{ root: classes.paper }} variant="outlined">
-                <Grid container={true} spacing={3}>
-                  <Grid item={true} xs={6}>
-                    <div className={classes.count}>
-                      {n(totalFiles)}
-                    </div>
-                    <div className={classes.countText}>
-                      {t('Files will be indexed')}
-                    </div>
-                  </Grid>
-                  <Grid item={true} xs={6}>
-                    <div className={classes.count}>
-                      {b(dataToIndex)}
-                    </div>
-                    <div className={classes.countText}>
-                      {t('Of data will be indexed')}
-                    </div>
-                  </Grid>
-                </Grid>
-              </Paper>
+             <FileIndexingConfigurationAndImpact
+               totalFiles={totalFiles}
+               dataToIndex={dataToIndex}
+             />
             </Grid>
             <Grid item={true} xs={6} style={{ marginTop: 30 }}>
-              <Typography variant="h4" gutterBottom={true}>
-               {t('Indexing information')}
-              </Typography>
-              <Paper classes={{ root: classes.paper }} variant="outlined">
-                <Grid container={true} spacing={3}>
-                  <Grid item={true} xs={4}>
-                    {isStarted ? (
-                      <Button
-                        startIcon={<PauseOutlined />}
-                        aria-label="Pause"
-                        onClick={handlePause}
-                        size="large"
-                        color="warning"
-                        variant="contained"
-                        classes={{ root: classes.button }}
-                      >
-                        {t('Pause')}
-                      </Button>
-                    ) : (
-                      <Button
-                        startIcon={<PlayArrowOutlined />}
-                        aria-label="Start"
-                        onClick={handleStart}
-                        size="large"
-                        color="success"
-                        variant="contained"
-                        classes={{ root: classes.button }}
-                      >
-                        {t('Start')}
-                      </Button>
-                    )}
-                  </Grid>
-                  <Grid item={true} xs={4}>
-                    <div className={classes.count}>
-                      {(isStarted ? n(indexedFiles) : '-')} / {n(totalFiles)}
-                    </div>
-                    <div className={classes.countText}>
-                      {t('Files indexed')}
-                    </div>
-                  </Grid>
-                  <Grid item={true} xs={4}>
-                    <div className={classes.count}>
-                      {(isStarted ? n(volumeIndexed) : '-')}
-                    </div>
-                    <div className={classes.countText}>
-                      {t('Volume indexed')}
-                    </div>
-                  </Grid>
-                </Grid>
-              </Paper>
+             <FileIndexingConfigurationInformations
+               indexedFiles={indexedFiles}
+               totalFiles={totalFiles}
+               volumeIndexed={volumeIndexed}
+               isStarted={isStarted}
+               handleStart={handleStart}
+               handlePause={handlePause}
+             />
             </Grid>
           </Grid>
         )}
